Splits eetplanmodel.class.php
